requirements.txt
anyio==4.9.0
asgiref==3.8.1
click==8.2.1
Django==5.2.3
graphql-core==3.2.6
h11==0.16.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.1
strawberry-graphql==0.275.5
strawberry-graphql-django==0.62.0
typing_extensions==4.14.0
uvicorn==0.35.0
